backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
weasyprint>=60.0
